Move config protocol to base.py and add docstring
