+++ alibi_detect/base.py
@@ -3,7 +3,9 @@ import copy
 import json
 import numpy as np
 from typing import Dict, Any, Optional
+from typing_extensions import Protocol, runtime_checkable
 from alibi_detect.version import __version__, __config_spec__
+
 
 DEFAULT_META = {
     "name": None,
@@ -173,6 +175,27 @@ class DriftConfigMixin:
                 inputs[key] = None
 
         self.config.update(inputs)
+
+
+@runtime_checkable
+class ConfigurableDetector(Protocol):
+    """Type Protocol for detectors that have support for saving via config.
+
+    Used for typing save and load functionality in `alibi_detect.saving.saving.py`.
+
+    Note:
+        This exists to distinguish between detectors with and without support for config saving and loading. Once all
+        detector support this then this protocol will be removed.
+    """
+    meta: Dict
+
+    def get_config(self): ...
+
+    def from_config(self): ...
+
+    def _set_config(self): ...
+
+    def predict(self): ...
 
 
 class NumpyEncoder(json.JSONEncoder):
